Jenkinsfile: drop support for generating the jenkins xunit test report
Since this is not really required to fail the build when the tests fail because our build.ps test does that, we can drop this.

We also need to drop this because the jUnit test results generated by dotnet are not compatible with the newer jenkins xUnit plugin (see https://issues.jenkins.io/browse/JENKINS-52152).